requirements.txt
scikit-image==0.16.2
scipy==1.4.1
matplotlib==3.1.2
numpy==1.18.1
opencv-python==4.2.0.32
tensorflow-gpu==2.2.0
